+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.18.12] - 2026-03-01
 
 - fix: Serialize only JSONB columns when building query parameters
   - Previously all objects/arrays were JSON.stringify'd before being passed to pg, which broke native Postgres array columns (text[], int[], etc.)